+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Mapbox welcomes participation and contributions from everyone.
 ### master
 
 - Added intersection search support to MapboxGeocoding [#1074](https://github.com/mapbox/mapbox-java/pull/1074)
+- Added support for Turf polygonToLine method [#1075](https://github.com/mapbox/mapbox-java/pull/1075)
 
 ### v4.9.0-alpha.1 - September 4, 2019